moved_char_match_amount_to_utils
