+++ domTree.py
@@ -5,6 +5,7 @@
 
 
 from html_ import HTML
+from utils import char_match_amount
 # from collections import defaultdict
 from custom_packages.searchTree import SearchTree
 import re
@@ -434,13 +435,3 @@ class DomTree:
         
 
             
-
-
-def char_match_amount(a: str, b: str) -> int|bool:
-    match_amount: int = 0
-    for index in range(min(len(a), len(b))):
-        if a[index] == b[index]:
-            match_amount += 1
-        else:
-            return match_amount
-    return index+1
